+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * All ACL APIs - Create, Update, Read and Delete Token.
  * All Event APIs - Fire and list events.
  * All KeyValue APIs - Write, Read, Delete Value.
+ * All Session APIs - Create, Delete, Read, List Node Sessions, List Sessions & Renew.
  * All Snapshot APIs - Generate and Restore.
  * All Status APIs - Get leader and Raft peers.
  * All Transaction APIs - Create with strong verbs and strong operation types for KV and strong results.